requirements.txt
airium
Pillow